+++ defaults/renderer.go
@@ -7,10 +7,17 @@ import (
 	"github.com/volatiletech/authboss"
 )
 
+var (
+	jsonDefaultFailures = []string{authboss.DataErr, authboss.DataValidation}
+)
+
 // There is a separate package that does HTML Rendering authboss-renderer
 
-// JSONRenderer simply renders the data provided in JSON
+// JSONRenderer simply renders the data provided in JSON.
+// Known failure keys in the HTMLData can be passed in to force a
+// status: failure in the JSON when they appear.
 type JSONRenderer struct {
+	Failures []string
 }
 
 // Load is a no-op since json doesn't require any templates
@@ -19,7 +26,25 @@ func (JSONRenderer) Load(names ...string) error {
 }
 
 // Render the data
-func (JSONRenderer) Render(ctx context.Context, page string, data authboss.HTMLData) (output []byte, contentType string, err error) {
+func (j JSONRenderer) Render(ctx context.Context, page string, data authboss.HTMLData) (output []byte, contentType string, err error) {
+	if _, hasStatus := data["status"]; !hasStatus {
+		failures := j.Failures
+		if len(failures) == 0 {
+			failures = jsonDefaultFailures
+		}
+
+		status := "success"
+		for _, failure := range failures {
+			val, has := data[failure]
+			if has && val != nil {
+				status = "failure"
+				break
+			}
+		}
+
+		data["status"] = status
+	}
+
 	b, err := json.Marshal(data)
 	if err != nil {
 		return nil, "", err